bug: Fix so books rerender when search is closed
If a book was added to a shelf on the search page, this change
would not be reflected after clicking close search. Main page would not
show new book on shelf unless page was refreshed. Fixed by calling
history.push when user clicks on close search to ensure main page
rerendered books.