fix: Set results with no main page bookshelf to 'none'
Prior to fix, results returned from BooksAPI.search had their shelf
state synced only if they had a bookshelf on the main page. The results
without a bookshelf on the main page were not modified since they were
assumed to have a default shelf of 'none'. Apparently not all result
books returned from search have a shelf state of 'none'. One of these
books would list a shelf on the search page when that book was not
present in a main page bookshelf. Fixed the issue by setting all result
books that are not present in a main page bookshelf to shelf of 'none'.